database relationship creato tra users e posts

+++ app/routers/post.py
@@ -21,6 +21,7 @@ def read_posts(session: Session = Depends(get_session)):
 
 @router.post("/", response_model=PostResponse,status_code=status.HTTP_201_CREATED)
 def create_post(post: Posts, session: Session = Depends(get_session),current_user: int = Depends(get_current_user)):
+    post.owner_id = current_user.id
     session.add(post)
     session.commit()
     session.refresh(post)
@@ -37,8 +38,11 @@ def read_post(post_id: int, session: Session = Depends(get_session)):
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_post(post_id: int, session: Session = Depends(get_session),current_user: int = Depends(get_current_user)):
     post = session.get(Posts, post_id)
+    post_owner_id = post.owner_id
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    if post_owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     session.delete(post)
     session.commit()
     return None
@@ -46,8 +50,11 @@ def delete_post(post_id: int, session: Session = Depends(get_session),current_us
 @router.put("/{post_id}", response_model=PostResponse,status_code=status.HTTP_200_OK)
 def update_post(post_id: int, post: Posts, session: Session = Depends(get_session),current_user: int = Depends(get_current_user)):
     db_post = session.get(Posts, post_id)
-    if not db_post:
+    post_owner_id = db_post.owner_id
+    if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    if post_owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     db_post.title = post.title
     db_post.content = post.content
     db_post.published = post.published
